+++ sdk/communication/communication-call-automation/review/communication-call-automation.api.md
@@ -606,9 +606,9 @@ export interface MuteParticipantResult {
 
 // @public (undocumented)
 export interface OutStreamingData {
-    AudioData?: AudioData;
-    Kind: MediaKind;
-    StopAudio?: StopAudio;
+    audioData?: AudioData;
+    kind: MediaKind;
+    stopAudio?: StopAudio;
 }
 
 // @public
@@ -1273,7 +1273,28 @@ export interface StopTranscriptionOptions extends OperationOptions {
 }
 
 // @public
-export function streamingData(packetData: string | ArrayBuffer): TranscriptionMetadata | TranscriptionData | AudioData | AudioMetadata;
+export class StreamingData {
+    static getStopAudioForOutbound(): string;
+    static getStreamingDataForOutbound(data: string): string;
+    // (undocumented)
+    static getStreamingKind(): StreamingDataKind;
+    static parse(data: string | ArrayBuffer): StreamingDataResult;
+}
+
+// @public (undocumented)
+export enum StreamingDataKind {
+    // (undocumented)
+    AudioData = "AudioData",
+    // (undocumented)
+    AudioMetadata = "AudioMetadata",
+    // (undocumented)
+    TranscriptionData = "TranscriptionData",
+    // (undocumented)
+    TranscriptionMetadata = "TranscriptionMetadata"
+}
+
+// @public (undocumented)
+export type StreamingDataResult = TranscriptionMetadata | TranscriptionData | AudioData | AudioMetadata;
 
 // @public
 export enum TextFormat {
