+++ sdk/communication/communication-call-automation/src/index.ts
@@ -5,10 +5,9 @@ export * from "./callConnection";
 export * from "./callMedia";
 export * from "./callRecording";
 export * from "./callAutomationEventParser";
-export * from "./streamingDataParser";
+export * from "./streamingData";
 export * from "./models/models";
 export * from "./models/options";
 export * from "./models/responses";
 export * from "./models/events";
-export * from "./models/transcription";
-export * from "./models/audio";
+export * from "./models/streaming";
